Moved browserstack library from utils to integrations dir
